Drop download_verify and gpg_recv
Concept of downloading gpg keys from keyservers were already a bad
idea, but after later events, it's definitely not a secure approach.

In the future, always download and verify gpg keys and add them to
the container repository. In addition, redesign download_verify, so
that the hash is given in the command line (and is thus part of the
container repository). This should ensure the file we download, is
the file we expect.

For reference, article about SKS Keyserver attack:
 https://gist.github.com/rjhansen/67ab921ffb4084c865b3618d6955275f