Set error and empty feedback file if it is too large
If the grading container tries to send a massive feedback dump
back to the MOOC-Grader, the /container-post view fails to
receive it due to settings.DATA_UPLOAD_MAX_MEMORY_SIZE. This
fix sets the feedback_size_error field to true and empties the
feedback file when it is too large. MOOC-Grader can then detect
these cases and provide informative feedback to the students.

Part of apluslms/mooc-grader#146